added reward functions in separate package
