+++ othello/game_logic/agents/random_agent.py
@@ -1,0 +1,21 @@
+import random
+
+from game_logic.agents.agent import Agent
+from game_logic.board import Board
+from utils.color import Color
+
+
+class RandomAgent(Agent):
+	def __init__(self, color: Color, immediate_reward_function):
+		super().__init__(color, immediate_reward_function)
+		self.name: str = 'Random'
+		self.immediate_reward_function = immediate_reward_function
+
+	def __str__(self):
+		return f'{self.name}{super().__str__()}'
+
+	def get_next_action(self, board: Board, legal_directions: dict) -> tuple:
+		location: tuple = random.choice(list(legal_directions.keys()))
+		legal_directions: list = legal_directions[location]
+
+		return location, legal_directions
