+++ othello/main.py
@@ -1,11 +1,11 @@
-from agent import Agent
-from board import Board
-from color import Color
+from game_logic.agents.agent import Agent
+from utils.color import Color
 from colorama import init
 from termcolor import colored
 
-from game import Game
-from random_agent import RandomAgent
+from game_logic.game import Game
+from game_logic.agents.random_agent import RandomAgent
+from utils.intermediate_reward_functions.difference_with_prev_board import difference_with_prev_board
 
 
 def main():
@@ -14,8 +14,8 @@ def main():
 
 	# check global variables
 	assert 1 <= num_episodes <= 10000, f'Invalid number of episodes: num_episodes should be between 1 and 10000, but got {num_episodes}'
-	assert black.score == 0 and black.num_games_won == 0, f'Invalid black agent'
-	assert white.score == 0 and white.num_games_won == 0, f'Invalid white agent'
+	assert black.num_games_won == 0, f'Invalid black agent'
+	assert white.num_games_won == 0, f'Invalid white agent'
 
 	print(f'\nPlayers:\n\t{black}\n\t{white}\n')
 
@@ -27,17 +27,16 @@ def main():
 
 	print()
 
-	assert 0 == black.score + white.score, 'The scores were miscalculated'
 	ties: int = num_episodes - black.num_games_won - white.num_games_won
-	if black.score > white.score:
+	if black.num_games_won > white.num_games_won:
 		print(colored(
 			f'BLACK {black.num_games_won:>5}/{num_episodes:>5} ({black.num_games_won:>5}|{white.num_games_won:>5}|{ties:>5})',
 			'red'))
-	elif black.score < white.score:
+	elif black.num_games_won < white.num_games_won:
 		print(colored(
 			f'WHITE {white.num_games_won:>5}/{num_episodes:>5} ({black.num_games_won:>5}|{white.num_games_won:>5}|{ties:>5})',
 			'green'))
-	elif black.score == white.score:
+	elif black.num_games_won == white.num_games_won:
 		print(colored(
 			f'DRAW  {black.num_games_won:>5}/{num_episodes:>5} ({black.num_games_won:>5}|{white.num_games_won:>5}|{ties:>5})',
 			'cyan'))
@@ -49,8 +48,8 @@ def main():
 
 if __name__ == "__main__":
 	# initialize global variables
-	black: Agent = RandomAgent(Color.BLACK)  # the black agent
-	white: Agent = RandomAgent(Color.WHITE)  # the white agent
+	black: Agent = RandomAgent(Color.BLACK, difference_with_prev_board)  # the black agent
+	white: Agent = RandomAgent(Color.WHITE, difference_with_prev_board)  # the white agent
 	board_size: int = 6  # the size of the board e.g. 8x8
 	num_episodes: int = 100  # the number of episodes e.g. 100
 	verbose: bool = False  # wetter or not to print intermediate steps
