+++ othello/utils/color.py
