+++ othello/utils/intermediate_reward_functions/difference_between_players.py
@@ -1,0 +1,5 @@
+from game_logic.board import Board
+
+
+def difference_between_players(self, board: Board, color_value: int) -> float:
+	raise NotImplementedError
